docs(consistency): align hybrid weights, rename JEPA to query_expander
Two consistency cleanups from docs/handovers/v0.1.3.1-hotfix.md (B2,
B3) — both factual fixes that align documentation with the code.

## B2 — README hybrid search weights

README L83 architecture diagram said
"Vector(60%) + BM25(20%) + keyword(20%)" — a 3-way decomposition that
sums to 100% but misses the 4th signal in the actual code.

The real weights at core/retrieval_engine.py::hybrid_search L146:

    score = 0.6*v + 0.2*b + 0.1*k + 0.1*n

That's a 4-way split (vector / BM25 / keyword / name). README now
matches: "Vector(60%) + BM25(20%) + keyword(10%) + name(10%)".

calculate_confidence at the bottom of the same file uses a different
distribution (0.5 / 0.2 / 0.2 / 0.1, also summing to 1.0). The two
functions have different intent — ranking vs answer-confidence — so
they intentionally diverge. A docstring on calculate_confidence now
spells that out so future readers don't assume one is a typo of the
other.

## B3 — core/jepa_adapter.py → core/query_expander.py

The original module's own docstring made it clear it
"❌ LLM 호출 금지 / ❌ Graph 접근 금지" — i.e. it never did anything
JEPA-related. It's a Korean keyword synonym dictionary plus a
stopword filter. Naming it "JEPA" (Joint-Embedding Predictive
Architecture, LeCun) misleads readers from the ML research side.

Changes:

- core/query_expander.py (new) — copy of the old module with:
  * docstring rewritten to state what it is and isn't
  * constants renamed: JEPA_TOKEN_HARD_LIMIT → TOKEN_HARD_LIMIT,
    JEPA_TIMEOUT_SEC → TIMEOUT_SEC (legacy aliases retained on the
    same module for backward compat through this minor)
  * print prefix [JEPA] → [QUERY-EXPAND]
  * log step prefix jepa.* → query_expand.*
- core/jepa_adapter.py — deprecation shim. Re-exports everything from
  core.query_expander and emits a DeprecationWarning on import. Plan:
  remove in v0.3.
- core/reasoning_engine.py L594-600 comment updated
- core/orchestrator.py L85, L169 comments updated
- tools/code/code_editor.py PROTECTED_FILES gains query_expander.py
  while keeping jepa_adapter.py (the shim) protected

Verification:

- import core.query_expander → expand("AI 연구 기관은?") returns
  expanded query with [QUERY-EXPAND] log prefix
- import core.jepa_adapter → succeeds + emits DeprecationWarning;
  re-exported expand is the same function object
- import core.reasoning_engine, core.orchestrator,
  tools.code.code_editor — all succeed
- TOKEN_HARD_LIMIT == JEPA_TOKEN_HARD_LIMIT == 50
